move custom components under components/application
We'll have our components organised in two levels:
- application (for custom components)
- ui (for ui components)
