+++ src/components/application/filters/QuickFilters/QuickFilters.tsx
